STYLE: Replace integer literals which are cast to bool.
Finds and replaces integer literals which are cast to bool.

SRCDIR=/Users/johnsonhj/src/vxl #My local SRC
BLDDIR=/Users/johnsonhj/src/vxl-bld/ #My local BLD

cd /Users/johnsonhj/src/vxl-bld/
run-clang-tidy.py -checks=-*,modernize-use-bool-literals  -header-filter=.* -fix

+++ contrib/brl/bseg/boxm2/boxm2_scene.cxx
@@ -605,7 +605,7 @@ vgl_vector_3d<unsigned int>  boxm2_scene::scene_dimensions() const
   }
   max_i++; max_j++; max_k++;
 
-  return {(max_i-min_i),(max_j - min_j),(max_k-min_k)};
+  return {static_cast<unsigned int>((max_i-min_i)),static_cast<unsigned int>((max_j - min_j)),static_cast<unsigned int>((max_k-min_k))};
 }
 
 //: gets the smallest block index
